backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
google-genai
pydantic
email-validator
